# Optimize _geometric_adstock for 99% function speedup, reducing total model train time by ~40% - Python version