## Project Robyn

### Summary  
This PR optimizes `_geometric_adstock`, the main performance bottleneck, reducing its execution time by **~99%** and improving the **total model train time by ~40%** on Python version.

### Fixes  
This PR does not address an existing GitHub issue but significantly improves performance.

### Type of change  
- **fix**: Performance optimization (non-breaking change that improves efficiency)
- **test**: Added unit test to validate correctness  

### Motivation & Context  
- `_geometric_adstock` previously accounted for **39% of the total model train time**, significantly slowing down model execution.  
- This function was rewritten using `scipy.signal.lfilter`, replacing a Python loop with a **highly efficient SciPy-based implementation (built on NumPy)**.  
- This change **improves execution time from 108.9s → 0.495s** (~99% speedup), leading to an **overall app model train time reduction of ~40%**.  

### **Performance Benchmark** (tutorial1.ipynb, Trial=1, iterations=2000)
**Before Optimization:**  
- `_geometric_adstock`: **108.9s** (39% of total model train time)  
- **Total model train time:** **277.9s**  

**After Optimization:**  
- `_geometric_adstock`: **0.495s** (**99% faster**)  
- **Total model train time:** **166.7s** (**40% faster**)